backend/requirements.txt
flask==3.0.3
requests==2.32.3
kafka-python==1.4.7
async-timeout==5.0.1
fastjsonschema==2.16.2
geojson==2.5.0
json5==0.9.25
jsonpointer==2.1
jsonschema==4.23.0
jsonschema-specifications==2023.7.1
orjson==3.10.7
python-json-logger==2.0.7
python-lsp-jsonrpc==1.1.2
ujson==5.10.0
threadpoolctl==3.5.0